showingPost function was created in home.js file
