+++ src/home.js
@@ -1,3 +1,6 @@
+import { db } from './firebase.js';
+
+
 const createPost = `
 <div>
   <div id="newPost">
@@ -13,9 +16,10 @@ const createPost = `
 </div>
 `;
 
+// Renderiza la tarjeta de la publicación en el muro.
 export const renderPost = (param) => `
 <div>
-  <div id="cardContainer" class= "link">
+  <div id="cardContainer" id ="link">
     <h2 id="cardTitle">${param.Title}</h2><img id="authorAvatar" class="icons" src="resources/user.png" alt="authorAvatar">
     <h3 id"cardSubtitle">${param.Subtitle}</h3>
     <p id="renderBody">${param.Body}</p>
@@ -23,18 +27,17 @@ export const renderPost = (param) => `
     </div>
 </div>
 `;
-// Renderiza la tarjeta de la publicación en el muro.
-export const createCard = (post) => {
-  // eslint-disable-next-line no-unused-expressions
-  `
-    <div>
-        <div id="cardContainer">
-            <h2 id="cardTitle">${post.titulo}</h2> <img id="authorAvatar" class="icons" src="resources/user.png" alt="authorAvatar">
-            <h3 id="cardSubtitle">${post.descripcion}</h3>
-            <p>${post.texto}</p>
-            <img id="readingTime" class="icons" src="resources/clock.png" alt="readingTime">
-        </div>
-  `;
+
+const showingPost = (postContainer) => {
+  db.collection('newPost')// .orderBy('fecha')
+    .onSnapshot((querySnapshot) => {
+      querySnapshot.forEach((doc) => {
+        const dataBase = doc.data();
+        // eslint-disable-next-line no-param-reassign
+        postContainer.innerHTML += renderPost(dataBase);
+        console.log(dataBase);
+      });
+    });
 };
 
 export const home = (container) => {
@@ -52,5 +55,7 @@ export const home = (container) => {
     <div id="container2"></div>
   `;
 
+  // eslint-disable-next-line no-param-reassign
   container.innerHTML = html;
+  showingPost(container);
 };
